requirements-desktop.txt
# Desktop-specific dependencies for pywebview
# Install with: pip install -r requirements-desktop.txt

# Cross-platform webview wrapper
pywebview>=4.0.0

# Platform-specific dependencies are handled automatically by pywebview:
# - Windows: pythonnet, pywin32
# - macOS: pyobjc
# - Linux: PyGObject (requires system packages: libgtk-3-dev, libwebkit2gtk-4.0-dev)

# Build tools (for creating executables)
# pyinstaller>=6.0.0  # Uncomment if building locally
